# src/main/resources/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Verdan Skoleadministrasjon API",
    "description": "REST API for Verdan skoleadministrasjonssystem. Støtter multi-institusjon med rollebasert tilgangskontroll (SUPER_ADMIN, INSTITUTION_ADMIN, TEACHER, STUDENT).\n\nAlle endepunkter (unntatt `/api/login`, `/api/auth/refresh` og `/api/health`) krever JWT Bearer-token i `Authorization`-headeren.\n\nAccess-tokens utløper etter 15 minutter. Bruk `/api/auth/refresh` for å fornye.",
    "version": "2.0.0",
    "contact": {
      "name": "Verdan Dev Team (Gruppe 2)"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Lokal utvikling"
    },
    {
      "url": "http://localhost:8081",
      "description": "Docker"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateUserRequest": {
        "type": "object",
        "required": ["username", "password", "role"],
        "properties": {
          "username": { "type": "string", "example": "newuser" },
          "password": { "type": "string", "example": "pass1234" },
          "role": { "type": "string", "enum": ["SUPER_ADMIN", "INSTITUTION_ADMIN", "TEACHER", "STUDENT"] },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List all users (ADMIN only)",
        "parameters": [
          { "name": "role", "in": "query", "schema": { "type": "string", "enum": ["SUPER_ADMIN", "INSTITUTION_ADMIN", "TEACHER", "STUDENT"] }, "description": "Filter by role" }
        ],
        "responses": {
          "200": { "description": "List of users" },
          "403": { "description": "Insufficient permissions" }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create a new user (ADMIN only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Validation error" },
          "409": { "description": "Username already exists" }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "User found" },
          "403": { "description": "Students can only view own profile" },
          "404": { "description": "User not found" }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update user (ADMIN only)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "User updated" },
          "404": { "description": "User not found" }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user (ADMIN only)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "204": { "description": "User deleted" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/users/import": {
      "post": {
        "tags": ["Users"],
        "summary": "Batch import users from Excel/CSV (ADMIN)",
        "requestBody": {
          "required": true,
          "content": { "multipart/form-data": { "schema": { "type": "object", "properties": { "file": { "type": "string", "format": "binary" } } } } }
        },
        "responses": {
          "200": { "description": "Import results with created/failed counts" },
          "400": { "description": "Invalid file format" }
        }
      }
    },
    "/api/users/batch-delete": {
      "post": {
        "tags": ["Users"],
        "summary": "Batch delete multiple users (ADMIN)",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "ids": { "type": "array", "items": { "type": "integer" } } } } } }
        },
        "responses": {
          "200": { "description": "Deleted count" },
          "403": { "description": "Cannot delete super admins" }
        }
      }
    },
    "/api/users/transfer": {
      "post": {
        "tags": ["Users"],
        "summary": "Batch transfer existing students to this institution via CSV/Excel (VGS ADMIN)",
        "description": "Reads a CSV/Excel file with columns: epost (or brukernavn), linje, trinn. Finds existing students by email/username and moves them to the calling admin's institution. Students keep their credentials, grades, and history. Uses all-or-nothing validation.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "CSV or Excel file with columns: epost, linje, trinn"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transfer results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transferred": { "type": "integer", "description": "Number of students successfully transferred" },
                        "skipped": { "type": "integer" },
                        "total": { "type": "integer", "description": "Total rows in the file" },
                        "errors": { "type": "array", "items": { "type": "string" }, "description": "Validation errors (if any, entire batch is rejected)" },
                        "transferredStudents": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "username": { "type": "string" },
                              "fullName": { "type": "string" },
                              "program": { "type": "string", "description": "Target program/linje name" },
                              "yearLevel": { "type": "string", "description": "VG1, VG2, or VG3" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "No file uploaded" },
          "500": { "description": "File parse error" }
        }
      }
    },
    "/api/users/{id}/transfer-history": {
      "get": {
        "tags": ["Users"],
        "summary": "Get grades and attendance from student's previous institution (ADMIN)",
        "description": "Returns the transfer history for a student who was transferred from another institution. Includes grades and attendance records from the previous school.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Transfer history with grades and attendance from previous institution" },
          "403": { "description": "Insufficient permissions" }
        }
      }
    },
...
  },
...
}